\name{trim32}
\alias{trim32}

\title{Gene expression levels of TRIM32}
\description{ Gene expression levels of TRIM32 from the microarray study by Scheetz et al. (2006). This data is also available in the \code{flare} package.
 }
\usage{trim32}
\format{
\describe{
    \item{\code{trim32}:}{is a response vector with 120 entries.}
  }
}

\details{The data set contains gene expression levels for 120 samples of TRIM32. It comes from the microarray experiments of mammalianeye tissue samples of Scheetz et al. (2006).
}
\source{ 
https://cran.r-project.org/web/packages/flare/
}
\references{
Scheetz, T. E., Kim K.-Y. A., Swiderski R. E., Philp A. R., Braun, T. A., Knudtson, K. L., Dorrance, A. M., DiBona, G. F., Huang, J., Casavant, T. L., Sheffield, V. C., and Stone, E. M. (2006). "Regulation of gene expression in the mammalian eye and its relevance to eye disease." \emph{Proceedings of the National Academy of Sciences}, \bold{103}(39):14429-14434.
}